refactor: add SelectionControls component
